Added a custom  hook to withErrorHandler component
