+++ build-a-burger/src/components/UI/Modal/Modal.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import classes from "./Modal.module.css";
 import Aux from "../../../hoc/Aux/Aux";
 import Backdrop from "../Backdrop/Backdrop";
