+++ build-a-burger/src/hoc/withErrorHandler/withErrorHandler.js
@@ -1,37 +1,15 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import Modal from "../../components/UI/Modal/Modal";
 import Aux from "../Aux/Aux";
+import useHttpErrorHandler from "../../containers/hooks/http-error-handler";
 
 const withErrorHandler = (WrappedComponent, axios) => {
   return (props) => {
-    const [error, setError] = useState(null);
-
-    const reqInterceptor = axios.interceptors.request.use((req) => {
-      setError(null);
-      return req;
-    });
-    const resInterceptor = axios.interceptors.response.use(
-      (res) => res,
-      (err) => {
-        setError(err);
-      }
-    );
-
-    //Clean up
-    useEffect(() => {
-      return () => {
-        axios.interceptors.request.eject(reqInterceptor);
-        axios.interceptors.response.eject(resInterceptor);
-      };
-    }, [reqInterceptor, resInterceptor]);
-
-    const errorConfirmedHandler = () => {
-      setError(null);
-    };
+    const [error, clearError] = useHttpErrorHandler(axios);
 
     return (
       <Aux>
-        <Modal show={error} modalClosed={errorConfirmedHandler}>
+        <Modal show={error} modalClosed={clearError}>
           {error ? error.message : null}
         </Modal>
         <WrappedComponent {...props} />
